add dropped attribute count to SpanData, Link and Event

+++ api/src/main/java/io/opentelemetry/trace/Link.java
@@ -51,5 +51,5 @@ public interface Link {
    *
    * @return the number of dropped attributes.
    */
-  int getDroppedAttributeCount();
+  long getDroppedAttributeCount();
 }
